// app/layout/models/store-fuel-grade-tax.model.ts
export class StoreFuelGradeTax {
  storeFuelGradeTaxID: number;
  storeFuelGradeID: number;
  fuelTaxDescription: string;
  fuelTaxRate: number;
  isApplyPrice: boolean;
  departmentID: number;
  storeFuelGradeNo: number;
  fuelGradeID: number;
  fuelGradeName: string;
  storeLocationID: number;
  storeName: string;
}